Jogo funcionando, só precisa atualizar as funções das peças especiais

+++ board.py
@@ -18,8 +18,8 @@ class Board(pygame.sprite.Sprite):
             self.info, self.map = None, None
 
         # Inicializa os mapas
-        self._create_level()
         self._set_info()
+        self._create_level()
 
     # Adiciona o modo de jogo
     def _set_mode(self):
@@ -49,6 +49,7 @@ class Board(pygame.sprite.Sprite):
         info = self.info
         self.points = 0
 
+        self.types = 3
         if info is not None:
             self.moves = info[0]
             self.w_points = info[1]
@@ -92,9 +93,8 @@ class Board(pygame.sprite.Sprite):
 
                         if t != t_aux:
                             break
-
-                        if index == 2:
-                            return True 
+                    else:
+                        return True
 
                 # Se for maior, não precisa verificar
                 if j + 2 < 9:
@@ -111,9 +111,8 @@ class Board(pygame.sprite.Sprite):
 
                         if t != t_aux:
                             break
-
-                        if index == 2:
-                            return True
+                    else:
+                        return True
         
         return False
 
@@ -132,12 +131,12 @@ class Board(pygame.sprite.Sprite):
                         if p == 0:
                             self.level[j][i] = piece.Block(j,i)
                         elif p == 1:
-                            self.level[j][i] = piece.Simple(j,i,6)
+                            self.level[j][i] = piece.Simple(j,i,self.types)
                         elif p == 2:
                             if first:
-                                self.level[j][i] = piece.Protection(j,i,6)
+                                self.level[j][i] = piece.Protection(j,i,self.types)
                             else:
-                                self.level[j][i] = piece.Simple(j,i,6)
+                                self.level[j][i] = piece.Simple(j,i,self.types)
                         elif p == 3:
                             self.level[j][i] = piece.Objective(j,i)
                         
@@ -149,7 +148,7 @@ class Board(pygame.sprite.Sprite):
                 for y in range(9):
                     check = True
                     while(check):
-                        self.level[x][y] = piece.Simple(x,y,6)
+                        self.level[x][y] = piece.Simple(x,y,self.types)
                         check = self._check_board(self.level)
     
     # Retorna uma lista com todas as peças
@@ -177,6 +176,8 @@ class Board(pygame.sprite.Sprite):
         # Faz uma copia do nivel e troca as posições
         self.level[p1.x][p1.y] = p1
         self.level[p2.x][p2.y] = p2
+        p1.update_rect()
+        p2.update_rect()
 
     # Checar todas as combinações com essa peça
     def _check_combinations(self, p):
@@ -184,7 +185,7 @@ class Board(pygame.sprite.Sprite):
         # Checa a sequencia a esquerda da peça
         x = p.x -1
         left = 0
-        while x > 0:
+        while x >= 0:
             
             # Verifica se é o mesmo tipo, caso ocorra algum
             # erro, é uma classe sem "tipo", então pula fora
@@ -218,7 +219,7 @@ class Board(pygame.sprite.Sprite):
         # Checa a sequência a cima da peça
         y = p.y -1
         top = 0
-        while y > 0:
+        while y >= 0:
             
             try:
                 if self.level[p.x][y].type == p.type:
@@ -268,7 +269,7 @@ class Board(pygame.sprite.Sprite):
             delete = True
             for i in range(p.x - left, p.x):
                 self.level[i][p.y] = None
-            for i in range(p.x +1, p.x + right):
+            for i in range(p.x +1, p.x + right + 1):
                 self.level[i][p.y] = None
         
         # Verifica a Coluna
@@ -276,7 +277,7 @@ class Board(pygame.sprite.Sprite):
             delete = True
             for j in range(p.y - top, p.y):
                 self.level[p.x][j] = None
-            for j in range(p.y + 1, p.y + bottom):
+            for j in range(p.y + 1, p.y + bottom + 1):
                 self.level[p.x][j] = None
         
         # Se a peça foi deletada, exclui ela
@@ -288,19 +289,30 @@ class Board(pygame.sprite.Sprite):
                 self.level[p.x][p.y] = new_piece
             else:
                 self.level[p.x][p.y] = None
+        
+        print(
+            "Left: %d" % left,
+            "Right: %d" % right,
+            "Top: %d" % top,
+            "Bottom: %d" % bottom
+        )
 
     # Destroi as peças do campo
     def _destroy_pieces(self, pieces):
         
         # Checa a combinação dessas duas peças
+        print("Começou a destruição")
         for p in pieces:
+            print("(%d,%d)" % (p.x, p.y))
             self._check_combinations(p)
-        
+            print("")
+        print("Acabou a destruição!")
+
         # Percorre todas as colunas
-        for y in range(9):
+        for x in range(9):
             
             # Vai da posição mais baixa até ao topo
-            for x in reversed(range(9)):
+            for y in reversed(range(9)):
 
                 # Se a posição for None, é necessário
                 # buscar as posições     
@@ -309,20 +321,19 @@ class Board(pygame.sprite.Sprite):
 
                     # Percorre até encontrar uma peça diferente
                     # de None e diferente de Block
-                    for i in reversed(range(x)):
+                    for i in reversed(range(y)):
 
-                        p = self.level[i][y]
+                        p = self.level[x][i]
                         if (p is not None) and (type(p) is not piece.Block):
                             find = True
                             
                             # Atualiza com a posição nova
-                            print(type(p))
                             p.x = x
                             p.y = y
                             self.level[x][y] = p
 
                             # Retira a peça da posição antiga
-                            self.level[i][y] = None
+                            self.level[x][i] = None
 
                             # Se tiver combinações, destroi na localização
                             # dessa peça
@@ -340,27 +351,40 @@ class Board(pygame.sprite.Sprite):
                     if not find:
                         
                         # Começa da peça até o topo
-                        for i in reversed(range(x+1)):
+                        for i in reversed(range(y+1)):
                             
-                            # Adiciona a nova peça
-                            self.level[i][y] = piece.Simple(i,y,6)
+                            # Se não for uma proteção, adiciona a peça nova!
+                            if type(self.level[x][i]) is not piece.Block:
 
-                            # Se gerou uma nova possibilidade, destruir!
-                            if self._check_board(self.level):
-                                return self._destroy_pieces(
-                                    [self.level[i][y]])
+                                # Adiciona a nova peça
+                                self.level[x][i] = piece.Simple(x,i,self.types)
+
+                                # Se gerou uma nova possibilidade, destruir!
+                                if self._check_board(self.level):
+                                    return self._destroy_pieces(
+                                        [self.level[x][i]])
                         
                         # Já adicionou peças o suficiente,
                         # pode pular pra próxima posição
                         break
-        
-        print("ROBERTO!")
 
         # O Tabuleiro está completo novamente, agora
         # se não tiver movimentos possiveis, gera novamente
         if not self._has_moves():
             self._create_level(first=False)
+        
+        # Atualiza as posições
+        self._update_rects()
 
+    # Atualiza os rects das peças
+    def _update_rects(self):
+        for x in range(9):
+            for y in range(9):
+                p = self.level[x][y]
+                p.x = x
+                p.y = y
+                p.update_rect()
+    
     # Usado para verificar se o tabuleiro possui movimentos possiveis
     def _has_moves(self):
